47328 refactor(realtime): consolidate WebSocket payload types into types.ts
Move all WS envelope and data interfaces from wsPayloads.ts into types.ts,
introduce ERealtimeEnvelopeType enum and notification_created envelope shape.
